Migrate caramel top page to app router
